Safari-609.1.11.1/LayoutTests/fast/dom/MutationObserver/text-content-on-attr-expected.txt

This tests mutating the attribute value by the setters of textContent and nodeValue. WebKit should insert exactly one record

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


element.attributes.title.textContent = "new"; records = observer.takeRecords()
PASS records[0].attributeName is "title"
PASS records.length is 1

element.attributes.title.nodeValue = "world"; records = observer.takeRecords()
PASS records[0].attributeName is "title"
PASS records.length is 1
PASS successfullyParsed is true

TEST COMPLETE
